+++ lz4.go
@@ -6,8 +6,10 @@ package lz4
 import "C"
 
 import (
+	"bytes"
 	"encoding/binary"
 	"errors"
+	"fmt"
 	"io"
 	"reflect"
 	"unsafe"
@@ -15,7 +17,11 @@ import (
 
 const (
 	streamingBlockSize        = 1024 * 64
+	blockHeaderSize           = 4
 	boundedStreamingBlockSize = streamingBlockSize + streamingBlockSize/255 + 16
+
+	hugeStreamingBlockSize        = 1024 * 1024 * 5
+	boundedHugeStreamingBlockSize = hugeStreamingBlockSize + hugeStreamingBlockSize/255 + 16
 )
 
 // p gets a char pointer to the first byte of a []byte slice
@@ -178,6 +184,7 @@ type reader struct {
 	isLeft           bool
 }
 
+// DEPRECATED: Use NewDecompressReader instead.
 // NewReader creates a new io.ReadCloser.  Reads from the returned ReadCloser
 // read and decompress data from r.  It is the caller's responsibility to call
 // Close on the ReadCloser when done.  If this is not done, underlying objects
@@ -214,6 +221,7 @@ func (r *reader) Close() error {
 }
 
 // Read decompresses `compressionBuffer` into `dst`.
+// dst buffer must of at least streamingBlockSize bytes large
 func (r *reader) Read(dst []byte) (int, error) {
 	if len(dst) == 0 {
 		return 0, nil
@@ -255,7 +263,7 @@ func (r *reader) Read(dst []byte) (int, error) {
 	if decompressed < 0 {
 		return decompressed, errors.New("error decompressing")
 	}
-	// fmt.Println(hex.EncodeToString(ptr[:]))
+
 	mySlice := C.GoBytes(ptr, C.int(decompressed))
 	copySize := min(decompressed, len(dst))
 
@@ -290,4 +298,213 @@ func (r *reader) readFromPending(dst []byte) (int, error) {
 		r.pending = r.pending[copied:]
 	}
 	return copied, nil
+}
+
+// CompressReader reads input and creates an io.ReadCloser for reading
+// compressed output
+type CompressReader struct {
+	underlyingReader       io.Reader
+	compressionBuffer      [2]unsafe.Pointer
+	outputBuffer           *bytes.Reader
+	lz4Stream              *C.LZ4_stream_t
+	inpBufIndex            int
+	totalCompressedWritten int
+	compressedBuffer       unsafe.Pointer
+}
+
+// NewCompressReader creates a new io.ReadCloser.  Reads from the returned ReadCloser
+// read and compress data from r.  It is the caller's responsibility to call
+// Close on the ReadCloser when done.  If this is not done, underlying objects
+// in the lz4 library will not be freed.
+func NewCompressReader(r io.Reader) *CompressReader {
+	return &CompressReader{
+		compressionBuffer: [2]unsafe.Pointer{
+			C.malloc(hugeStreamingBlockSize),
+			C.malloc(hugeStreamingBlockSize),
+		},
+		lz4Stream:        C.LZ4_createStream(),
+		underlyingReader: r,
+		outputBuffer:     bytes.NewReader(nil),
+		compressedBuffer: C.malloc(boundedHugeStreamingBlockSize + blockHeaderSize),
+	}
+}
+
+// Read compresses data from the underlyingReader into dst.
+func (r *CompressReader) Read(dst []byte) (int, error) {
+	// try to consume from the buffer
+	n, _ := r.outputBuffer.Read(dst)
+	// ignoring err which can only be EOF in which case bytes read is 0
+	if n > 0 {
+		// if the buffer contains anything it's leftover from a previous call
+		return n, nil
+	}
+
+	// the buffer is empty, we are going to write into it so we reset it first
+	totalBlockSize := boundedHugeStreamingBlockSize + blockHeaderSize
+	inpPtr := r.nextInputBuffer()
+	outPtr := ptrToByteSlice(r.compressedBuffer, totalBlockSize, totalBlockSize)
+
+	bytesRead, err := io.ReadFull(r.underlyingReader, inpPtr)
+	if err == io.EOF {
+		// nothing left to read from the source
+		return 0, err
+	}
+	if err != nil && err != io.ErrUnexpectedEOF {
+		// ErrUnexpectedEOF occurs when some bytes are read but not all the bytes (n > 0)
+		return 0, fmt.Errorf("error reading source: %s", err)
+	}
+
+	// compress and write the data into compressedBuf, leaving space for the
+	// 4 byte header
+	written := int(C.LZ4_compress_fast_continue(
+		r.lz4Stream,
+		p(inpPtr),
+		p(outPtr[blockHeaderSize:]),
+		C.int(bytesRead),
+		C.int(boundedHugeStreamingBlockSize),
+		1))
+	if written <= 0 {
+		return 0, errors.New("error compressing")
+	}
+
+	// write "header" to the buffer for decompression at the first 4 bytes
+	binary.LittleEndian.PutUint32(outPtr[:blockHeaderSize], uint32(written))
+
+	// populate the buffer with our internal slice and consume from it
+	r.outputBuffer = bytes.NewReader(outPtr[:written+blockHeaderSize])
+	n, _ = r.outputBuffer.Read(dst)
+	// here we ignore any EOF because the buffer contains partial data only
+	// EOF will be communicated on the next call if the underlying Reader is exhausted
+
+	r.totalCompressedWritten += written + 4
+	return n, nil
+}
+
+func (r *CompressReader) nextInputBuffer() []byte {
+	r.inpBufIndex = (r.inpBufIndex + 1) % 2
+	return ptrToByteSlice(r.compressionBuffer[r.inpBufIndex], hugeStreamingBlockSize, hugeStreamingBlockSize)
+}
+
+// Close releases all the resources occupied by Reader.
+// r cannot be used after the release.
+func (r *CompressReader) Close() error {
+	if r.lz4Stream != nil {
+		C.LZ4_freeStream(r.lz4Stream)
+		r.lz4Stream = nil
+	}
+	C.free(r.compressionBuffer[0])
+	C.free(r.compressionBuffer[1])
+	C.free(r.compressedBuffer)
+	return nil
+}
+
+// DecompressReader is an io.ReadCloser that decompresses when read from.
+type DecompressReader struct {
+	lz4Stream           *C.LZ4_streamDecode_t
+	outputBuffer        *bytes.Reader
+	decompressionBuffer [2]unsafe.Pointer
+	underlyingReader    io.Reader
+	inpBufIndex         int
+	compressedBuffer    unsafe.Pointer
+}
+
+// NewDecompressReader creates a new io.ReadCloser. This function mirrors the
+// behavior of NewReader but provides better performance.
+// It is the caller's responsibility to call Close on the ReadCloser when done.
+// If this is not done, underlying objects in the lz4 library will not be freed.
+func NewDecompressReader(r io.Reader) io.ReadCloser {
+	return &DecompressReader{
+		lz4Stream:        C.LZ4_createStreamDecode(),
+		underlyingReader: r,
+		decompressionBuffer: [2]unsafe.Pointer{
+			// double buffer needs to use C.malloc to make sure the same memory address
+			// allocate buffers in go memory will fail randomly since GC may move the memory
+			C.malloc(hugeStreamingBlockSize),
+			C.malloc(hugeStreamingBlockSize),
+		},
+		outputBuffer:     bytes.NewReader(nil),
+		compressedBuffer: C.malloc(boundedHugeStreamingBlockSize),
+	}
+}
+
+// Read decompresses data from the underlying reader into `dst`.
+func (r *DecompressReader) Read(dst []byte) (int, error) {
+	// write data read from a previous call
+	n, _ := r.outputBuffer.Read(dst)
+	// ignoring err which can only be EOF in which case bytes read is 0
+	if n > 0 {
+		// if the buffer contains anything it's leftover from a previous call
+		return n, nil
+	}
+
+	compressedBlockSize, err := r.readSize(r.underlyingReader)
+	if err != nil {
+		return 0, err
+	}
+
+	inPtr := ptrToByteSlice(r.compressedBuffer, boundedHugeStreamingBlockSize, boundedHugeStreamingBlockSize)
+	outPtr := r.nextDecompressionBuffer()
+
+	// read the compressed blockSize from r.underlyingReader
+	_, err = io.ReadFull(r.underlyingReader, inPtr[:compressedBlockSize])
+	if err != nil {
+		return 0, err
+	}
+
+	decompressed := int(C.LZ4_decompress_safe_continue(
+		r.lz4Stream,
+		p(inPtr),
+		p(outPtr),
+		C.int(compressedBlockSize),
+		C.int(hugeStreamingBlockSize),
+	))
+
+	if decompressed < 0 {
+		return decompressed, errors.New("error decompressing")
+	}
+
+	// write the decompressed data to the output buffer
+	r.outputBuffer = bytes.NewReader(outPtr[:decompressed])
+	// read as much as we can into dst, ignoring any EOF
+	n, _ = r.outputBuffer.Read(dst)
+
+	return n, nil
+}
+
+// Close releases all the resources occupied by r.
+// r cannot be used after the release.
+func (r *DecompressReader) Close() error {
+	if r.lz4Stream != nil {
+		C.LZ4_freeStreamDecode(r.lz4Stream)
+		r.lz4Stream = nil
+	}
+
+	C.free(r.decompressionBuffer[0])
+	C.free(r.decompressionBuffer[1])
+	C.free(r.compressedBuffer)
+	return nil
+}
+
+func (r *DecompressReader) nextDecompressionBuffer() []byte {
+	r.inpBufIndex = (r.inpBufIndex + 1) % 2
+	return ptrToByteSlice(r.decompressionBuffer[r.inpBufIndex], hugeStreamingBlockSize, hugeStreamingBlockSize)
+}
+
+// read the 4-byte little endian size from the head of each stream compressed block
+func (r *DecompressReader) readSize(rdr io.Reader) (int, error) {
+	var temp [blockHeaderSize]byte
+	_, err := io.ReadFull(rdr, temp[:])
+	if err != nil {
+		return 0, err
+	}
+	return int(binary.LittleEndian.Uint32(temp[:])), nil
+}
+
+func ptrToByteSlice(dataPtr unsafe.Pointer, _len, _cap int) []byte {
+	tmpSlice := reflect.SliceHeader{
+		Data: uintptr(dataPtr),
+		Len:  _len,
+		Cap:  _cap,
+	}
+	return *(*[]byte)(unsafe.Pointer(&tmpSlice))
 }